requirements.txt
# Core LLM framework
llama-index-core

# Vector store
chromadb
llama-index-vector-stores-chroma

# Ollama integrations
llama-index-embeddings-ollama
llama-index-llms-ollama
